rapikan ke model
